+++ packages/edit-site/src/components/sidebar-navigation-screen-navigation-menus/navigation-menu-content.js
@@ -5,7 +5,7 @@ import {
 } from '@wordpress/block-editor';
 import { useDispatch, useSelect } from '@wordpress/data';
 import { createBlock } from '@wordpress/blocks';
-import { useCallback, useState } from '@wordpress/element';
+import { useCallback } from '@wordpress/element';
 import { store as coreStore } from '@wordpress/core-data';
 import { privateApis as blockLibraryPrivateApis } from '@wordpress/block-library';
 import { unlock } from '../../lock-unlock';
@@ -31,8 +31,6 @@ const PAGES_QUERY = [
 ];
 
 export default function NavigationMenuContent( { rootClientId } ) {
-	const [ editingBlock, setEditingBlock ] = useState( null );
-
 	const { listViewRootClientId, isLoading } = useSelect(
 		( select ) => {
 			const {
@@ -89,33 +87,6 @@ export default function NavigationMenuContent( { rootClientId } ) {
 		[ __unstableMarkNextChangeAsNotPersistent, replaceBlock ]
 	);
 
-	const LeafMoreMenuWithEditingBlock = useCallback(
-		( props ) => {
-			return (
-				<LeafMoreMenu
-					{ ...props }
-					setEditingBlock={ setEditingBlock }
-				/>
-			);
-		},
-		[ setEditingBlock ]
-	);
-
-	const NavigationLinkUIWithEditingBlock = useCallback(
-		( { block, insertedBlock, setInsertedBlock } ) => {
-			return (
-				<NavigationLinkUI
-					block={ block }
-					insertedBlock={ insertedBlock }
-					setInsertedBlock={ setInsertedBlock }
-					editingBlock={ editingBlock }
-					setEditingBlock={ setEditingBlock }
-				/>
-			);
-		},
-		[ editingBlock ]
-	);
-
 	// The hidden block is needed because it makes block edit side effects trigger.
 	// For example a navigation page list load its items has an effect on edit to load its items.
 	return (
@@ -124,9 +95,9 @@ export default function NavigationMenuContent( { rootClientId } ) {
 				<PrivateListView
 					rootClientId={ listViewRootClientId }
 					onSelect={ offCanvasOnselect }
-					blockSettingsMenu={ LeafMoreMenuWithEditingBlock }
+					blockSettingsMenu={ LeafMoreMenu }
 					showAppender
-					additionalBlockContent={ NavigationLinkUIWithEditingBlock }
+					additionalBlockContent={ NavigationLinkUI }
 					isExpanded
 				/>
 			) }